graphql: Refactor `QueryBlockCache`
