+++ graphql/src/execution/mod.rs
@@ -5,6 +5,10 @@ mod query;
 /// Common trait for field resolvers used in the execution.
 mod resolver;
 
+use stable_hash::{crypto::SetHasher, StableHasher};
+
 pub use self::execution::*;
 pub use self::query::Query;
 pub use self::resolver::Resolver;
+
+type QueryHash = <SetHasher as StableHasher>::Out;
